+++ backend/canisters/group/impl/src/timer_job_types.rs
@@ -9,7 +9,6 @@ use constants::{DAY_IN_MS, MINUTE_IN_MS, NANOS_PER_MILLISECOND, SECOND_IN_MS};
 use ledger_utils::process_transaction;
 use serde::{Deserialize, Serialize};
 use tracing::error;
-use types::icrc1::Account;
 use types::{BlobReference, MessageId, MessageIndex, P2PSwapStatus, PendingCryptoTransaction, TimestampMillis, UserId};
 
 #[derive(Serialize, Deserialize, Clone)]
@@ -265,12 +264,11 @@ impl Job for FinalPrizePaymentsJob {
 
 impl Job for MakeTransferJob {
     fn execute(self) {
-        let sender = read_state(|state| Account::from(state.env.canister_id()));
         let pending = self.pending_transaction.clone();
-        ic_cdk::futures::spawn_migratory(make_transfer(pending, sender, self.attempt));
+        ic_cdk::futures::spawn_migratory(make_transfer(pending, self.attempt));
 
-        async fn make_transfer(mut pending_transaction: PendingCryptoTransaction, sender: Account, attempt: u32) {
-            if let Err(error) = process_transaction(pending_transaction.clone(), sender, true).await {
+        async fn make_transfer(mut pending_transaction: PendingCryptoTransaction, attempt: u32) {
+            if let Err(error) = process_transaction(pending_transaction.clone(), None, true).await {
                 error!(?error, "Transaction failed");
                 if attempt < 50 {
                     mutate_state(|state| {